Removing unecessary files.
The Makefile was necessary for sync'ing between the local code repository
and $HOME/vimfiles. However, turns out that the easier thing to do is just
to 'set rtp+=/path/to/local/depo/vimfiles' in .vimrc. Then there is no mess
with sync'ing etc. Dunno why I did not think of this before.