BugFix: Multiple people using the same machine
Fix a problem where if multiple people were using the server on the same
machine, the port would "already be in use" and the two users would keep
kicking each other off. This only works when showing the async window is
off. In the other case, a more elaborate mechanism needs to be a devised.