Fix bug with startup 'GDB is busy'
If somethign in the GDB startup takes more than about 0.4 seconds, then we
used to throw a warning 'GDB is busy. Try again later'. I moved the initial
read to the GDB prompt to the constructor of the TerminalServer method.
This ensures that the startGdbServer function blocks till the first prompt
has arrived.