Fix problem with initial connection
VimGdbClient was erroring out when GDB was first started. This was
happening because we returned from the server after only bind()'ing the
socket, before we did a listen(). However, the client immediately tried to
do a connect(). This caused an initial "connection refused" error which
manifested as a "GDB is busy. Please try again" immediately after starting
GDB. Somehow only seems to happen on my Linux Mint 12 VM.